add progress

+++ GroupWork/Models/Record.cs
@@ -9,8 +9,8 @@ namespace GroupWork.Models
     {
         public int RecordID { get; set; }
         public int UserID { get; set; }
-        public int Weight { get; set; }
-        public int Height { get; set; }
+        public float Weight { get; set; }
+        public float Height { get; set; }
         public DateTime InputDate { get; set; }
 
         public virtual User User { get; set; }
